refactor(react-ui): reorganize primitives/ and components/ by role
Aligns the directory split with the conceptual layer:
- primitives/ = foundational infrastructure (unstyled layout boxes + runtime context providers)
- components/ = themed structural components with semantic regions

Moves:
- ThemeProvider, DensityProvider, ElevationProvider: components/ → primitives/
  (runtime React contexts, not UI widgets; they pair with use*Context hooks)
- Panel, Column: primitives/ → components/
  (have semantic regions like Panel.Root/Toolbar/Content/Statusbar and registered themes;
   structurally closer to Card/Dialog than to Flex/Grid/Container)

External API surface is unchanged — everything flows through @dxos/react-ui's main barrel.
Only internal relative imports were updated.
