+++ packages/ui/react-ui/src/components/Clipboard/CopyButton.tsx
@@ -6,9 +6,9 @@ import React from 'react';
 
 import { mx, osTranslations } from '@dxos/ui-theme';
 
+import { useTranslation } from '../../primitives/ThemeProvider';
 import { Button, type ButtonProps, IconButton } from '../Button';
 import { Icon, type IconProps } from '../Icon';
-import { useTranslation } from '../ThemeProvider';
 import { type TooltipScopedProps, useTooltipContext } from '../Tooltip';
 import { useClipboard } from './ClipboardProvider';
