+++ packages/ui/react-ui/src/components/Column/Column.stories.tsx
@@ -71,7 +71,7 @@ const DefaultStory = () => {
 };
 
 const meta: Meta = {
-  title: 'ui/react-ui-core/primitives/Column',
+  title: 'ui/react-ui-core/components/Column',
   render: DefaultStory,
   decorators: [withTheme(), withLayout({ layout: 'column', classNames: 'w-[25rem]' })],
   parameters: {
