+++ packages/ui/react-ui/src/components/List/List.tsx
@@ -27,9 +27,9 @@ import {
 import { type Density } from '@dxos/ui-types';
 
 import { useDensityContext, useThemeContext } from '../../hooks';
+import { DensityProvider } from '../../primitives/DensityProvider';
 import { composable, composableProps } from '../../util';
 import { type ThemedClassName } from '../../util';
-import { DensityProvider } from '../DensityProvider';
 import { Icon } from '../Icon';
 import { ListDropIndicator } from './ListDropIndicator';
