+++ packages/ui/react-ui/src/components/index.ts
@@ -2,16 +2,13 @@
 // Copyright 2022 DXOS.org
 //
 
-export * from './DensityProvider';
-export * from './ElevationProvider';
-export * from './ThemeProvider';
-
 export * from './Avatars';
 export * from './Breadcrumb';
 export * from './Button';
 export * from './Card';
 export * from './Carousel';
 export * from './Clipboard';
+export * from './Column';
 export * from './Dialog';
 export * from './ErrorFallback';
 export * from './Focus';
@@ -24,14 +21,15 @@ export * from './Main';
 export * from './MediaPlayer';
 export * from './Menu';
 export * from './Message';
+export * from './Panel';
 export * from './Popover';
-export * from './Status';
 export * from './ScrollArea';
 export * from './ScrollContainer';
 export * from './Select';
 export * from './Separator';
 export * from './Skeleton';
 export * from './Splitter';
+export * from './Status';
 export * from './Tag';
 export * from './Toast';
 export * from './Toolbar';
