+++ packages/ui/react-ui/src/hooks/useDensityContext.ts
@@ -6,7 +6,7 @@ import { useContext } from 'react';
 
 import { type Density } from '@dxos/ui-types';
 
-import { DensityContext } from '../components';
+import { DensityContext } from '../primitives';
 
 export const useDensityContext = (densityProp?: Density): Density | undefined => {
   const { density } = useContext(DensityContext);
