+++ packages/ui/react-ui/src/hooks/useElevationContext.ts
@@ -6,7 +6,7 @@ import { useContext } from 'react';
 
 import { type Elevation } from '@dxos/ui-types';
 
-import { ElevationContext } from '../components';
+import { ElevationContext } from '../primitives';
 
 export const useElevationContext = (propsElevation?: Elevation) => {
   const { elevation } = useContext(ElevationContext);
