+++ packages/ui/react-ui/src/hooks/useTranslationsContext.ts
@@ -4,6 +4,6 @@
 
 import { useContext } from 'react';
 
-import { TranslationsContext } from '../components/ThemeProvider/TranslationsProvider';
+import { TranslationsContext } from '../primitives/ThemeProvider/TranslationsProvider';
 
 export const useTranslationsContext = () => useContext(TranslationsContext);
