+++ packages/ui/react-ui/src/primitives/index.ts
@@ -2,8 +2,10 @@
 // Copyright 2026 DXOS.org
 //
 
-export * from './Column';
+export * from './DensityProvider';
+export * from './ElevationProvider';
+export * from './ThemeProvider';
+
 export * from './Container';
 export * from './Flex';
 export * from './Grid';
-export * from './Panel';
