+++ packages/ui/react-ui/src/util/index.ts
@@ -4,6 +4,6 @@
 
 export type { ThemedClassName } from '@dxos/ui-types';
 
-export * from './hasIosKeyboard';
+export * from './mobile';
 export * from './slots';
 export * from './usePx';
