+++ packages/ui/react-ui/src/util/usePx.ts
@@ -13,13 +13,15 @@ const getDocumentElementFontSize = () => parseFloat(getComputedStyle(document.do
  * @returns The current pixel value equivalent of the rem input
  */
 // TODO(burdon): Change to array.
+// TODO(burdon): Integrate with theme provider.
 export const usePx = (rem: number): number => {
   const [fontSize, setFontSize] = useState(() => {
     if (typeof document !== 'undefined') {
       return getDocumentElementFontSize();
     }
 
-    return 16; // Default fallback for SSR
+    // Default fallback for SSR
+    return 16;
   });
 
   const updateFontSize = useCallback(() => {
